requirements.txt
django==1.6.6
django-xadmin==0.5.0
MySQL-python==1.2.5
django-crispy-forms==1.4.0
django-reversion==1.8.1